- BUG: Changed float parsing of marker positions in the BrapiMapImporter to nf.parse(pos).doubleValue().

+++ src/jhi/flapjack/io/brapi/BrapiMapImporter.java
@@ -53,7 +53,7 @@ public class BrapiMapImporter implements IMapImporter
 			// Each MapEntry represents a marker: its name, chromosome, and
 			// location on chromosome
 
-			float position = nf.parse(bm.getLocation()).floatValue();
+			double position = nf.parse(bm.getLocation()).doubleValue();
 			String chromosome = bm.getLinkageGroup();
 
 			Marker marker = new Marker(bm.getMarkerName(), position);
